backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
botocore==1.42.16
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.1
fastapi==0.127.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langsmith==0.5.0
markupsafe==3.0.3
memori==3.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.0
orjson==3.11.5
packaging==25.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pyfiglet==1.0.4
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
